# src/main/java/com/example/habitstracker/controllers/HabitApiImpl.java
package com.example.habitstracker.controllers;

import com.example.habitstracker.models.Habit;
import com.example.habitstracker.security.UserCredentials;
import com.example.habitstracker.services.HabitService;
import com.example.habitstracker.services.MapperService;
import com.example.openapi.api.HabitApi;
import com.example.openapi.dto.HabitDTO;
import com.example.openapi.dto.IdDTO;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.RestController;

/**
 * Контроллер для привычки
 */
@RestController
public class HabitApiImpl implements HabitApi
{
    private final HabitService habitService;
    private final MapperService mapperService;

    @Autowired
    public HabitApiImpl(HabitService habitService, MapperService mapperService)
    {
        this.habitService = habitService;
        this.mapperService = mapperService;
    }

    @Override
    public ResponseEntity<IdDTO> createHabit(HabitDTO habitDTO)
    {
        UserCredentials credentials = (UserCredentials) SecurityContextHolder.getContext().getAuthentication().getCredentials();
        Habit habit = new Habit();
        mapperService.transform(habitDTO, habit);
        habitService.addHabit(credentials, habit);
        return ResponseEntity.ok(new IdDTO().id(habit.getId()));
    }

    @Override
    public ResponseEntity<Void> deleteHabit(Long id)
    {
        habitService.deleteHabit(id);
        return ResponseEntity.ok().build();
    }

    @Override
    public ResponseEntity<HabitDTO> getHabit(Long id)
    {
        HabitDTO habitDTO = new HabitDTO();
        mapperService.transform(habitService.getHabit(id), habitDTO);
        return ResponseEntity.ok(habitDTO);
    }

    @Override
    public ResponseEntity<Void> updateHabit(Long id, HabitDTO habitDTO)
    {
        Habit habit = new Habit();
        mapperService.transform(habitDTO, habit);
        habitService.updateHabit(id, habit);
        return ResponseEntity.ok().build();
    }
}